+++ lib/setup.php
@@ -122,3 +122,18 @@ function update_theme_settings_defaults() {
 function get_theme_settings_defaults() {
 	return get_configuration_parameters( 'theme_default_settings' );
 }
+
+
+/**
+ * Get the project settings defaults.
+ *
+ * @since 1.0.5
+ *
+ * @return array
+ */
+function get_project_settings_defaults() {
+	return get_configuration_parameters( 'project_settings' );
+}
+
+
+